narrow solver populations to float32
The kernel is memory-bound, so halving the population arrays is worth more
than any arithmetic change. All math stays in float64; only storage narrows,
which drops the solver footprint from 11.5 to 6.6 MB.

Measured on a Pi 4 (four Cortex-A72), where the parallel path saturates the
memory bus: one core gains 6%, but four cores gain 29% (12.80 -> 9.09 ms per
step), and scaling improves from 1.94x to 2.65x. The win tracks memory
pressure, which is why it barely shows on one core and shows up fully once
every core is competing for the bus. macOS gains a little too (frame 2.4%,
serial 3.4%), so nothing regresses there.

Physics drift is at the float32 noise floor, about 2e-6 relative on the
forces, and Sep is exact since it is a ratio of face counts. precision_test
pins that against the float64 reference values.

+++ lbm/stream_equiv_test.go
@@ -61,10 +61,10 @@ func TestStreamMatchesReference(t *testing.T) {
 		s.applyBoundaries()
 		s.computeForce()
 
-		ref := make([][]float64, 9)
+		ref := make([][]float32, 9)
 		s.streamReference()
 		for i := range 9 {
-			ref[i] = append([]float64(nil), s.ftmp[i]...)
+			ref[i] = append([]float32(nil), s.ftmp[i]...)
 		}
 		s.stream()
 		for i := range 9 {
